Use face skin colours exclusively for software block devices (#183)
Currently the Face Skin colour range from the GNOME palette represent a
mixture of file systems and software block devices:
    JFS          - Face Skin Medium
    LVM2_PV      - Face Skin Dark
    NILFS2       - Face Skin Shadow
    LINUX_SWRAID - Dark Brown
    ATARAID      - Dark Brown

We are about to add recognition of bcache [1][2][3], another software
block device.  Reorganise the colour assignments so that Face Skin
colour range is exclusively used by types of software block devices and
assign JFS and NILFS2 new colours.
    Face Skin Light  (#EFE0CD) -
    Face Skin Medium (#E0C39E) - BCACHE         [New assignment]
    Face Skin Dark   (#B39169) - LVM2_PV
    Face Skin Shadow (#826647) - LINUX_SWRAID   [New assignment]
    Brown Dark       (#5A4733) - ATARAID

NILFS2 has flash friendly characteristics [4] so use Accent Red colours
along with F2FS.
    Accent Red      (#DF421E) - F2FS
    Accent Red Dark (#990000) - NILFS2   [New assignment]

Move JFS to a group with XFS and UFS.  The hue of the GNOME palette
Accent Yellow Dark colour, as used by UFS, was more Orange compared to
Accent Yellow and a bit too close to the Orange used by BTRFS.  So use
the hue of the GNOME Accent Yellow and extend the range and assign like
this.
    Accent Yellow             (#EED680) - XFS
    Accent Yellow Dark        (#D6B129) - JFS   [Updated hue.
                                                 New assignment]
    Accent Yellow Shadow      (#AA8F2C) - UFS   [New colour.
                                                 New assignment]
    Accent Yellow Dark Shadow (#826F2B) -       [New colour]

[1] bcache
    https://bcache.evilpiepirate.org/
[2] Linux kernel document: A block layer cache (bcache)
    https://www.kernel.org/doc/Documentation/bcache.txt
[3] The Linux kernel user's and administrator's guide > A block layer
    cache (bcache)
    https://www.kernel.org/doc/html/latest/admin-guide/bcache.html
[4] NILFS, Relative performance
    https://en.wikipedia.org/wiki/NILFS#Relative_performance

Closes #183 - Basic support for bcache