Fix crash when copying NTFS to starting beyond 2 TiB (#164)
Create test setup using a 4 TiB loop device:
    # truncate -s 4T /tmp/disk.img
    # losetup -f --show /tmp/disk.img
    /dev/loop0
Create 2 x 1 TiB partitions.  First at offset 1 MiB, second at offset
2 TiB:
    # sgdisk --new 1:2048:2147485696 --typecode 1:0700 /dev/loop0
    # sgdisk --new 2:4294967296:6442450944 --typecode 2:0700 /dev/loop0
    # partprobe /dev/loop0
Create NTFS file system in the first partition:
    # mkntfs -Q /dev/loop0p1

Then use GParted to copy the first NTFS partition into the second
partition.  GParted crashes:
    # gpartedbin /dev/loop0
    ...
    (gpartedbin:14660): glibmm-ERROR **: 20:39:01.191:
    unhandled exception (type std::exception) in signal handler:
    what: basic_string::_M_replace_aux

    Trace/breakpoint trap
    # echo $?
    133

Overview of what is happening is that GParted_Core::update_bootsector()
is attempting to set the Hidden Sectors [1] field in the NTFS Partition
Boot Sector (PBS) to the start sector of the newly copied /dev/loop0p2
partition.  But the sector number is greater than will fit in a 32-bit
unsigned integer, which the code doesn't handle.

Specifically the code prints the sector number as a hexadecimal number
into string 'hex'.  As the target partition starts at exactly 2 TiB then
hex="100000000" (9 hexadecimal digits long).  Next:
    hex.insert(0, 8 - hex.length(), '0');
is meant to pad the beginning of the 'hex' string with '0's to make the
string 8 character long.  But the string is already 9 character long so
8 - 9 is -1 which as unsigned integral type size_t [2] is 2^64-1.  So
insert() is trying to insert 18446744073709551615 '0's at the start of
the 'hex' string!  Hence the crash.

mkntfs refuses to accept an explicit partition start sector of 2^32 or
larger:
    # mkntfs -Q --partition-start 4294967296 /dev/loop0p2
    Invalid partition start sector.  Maximum is 4294967295 (2^32-1).
    # echo $?
    1

When mkntfs can't determine the drive geometry and partition offset, as
is the case on loop devices or the partition start sector is 2^32 or
larger, then mkntfs writes zero into Hidden Sectors:
    # mkntfs -Q /dev/loop0p1
    The partition start sector was not specified for /dev/loop0p1 and it could not be obtained automatically.  It has been set to 0.
    ...
    To boot from a device, Windows needs the 'partition start sector', ...
    Windows will not be able to boot from this device.
    Creating NTFS volume structures.
    mkntfs completed successfully. Have a nice day.
    # echo $?
    0
    # hexdump -C /dev/loop0p1 | head -2
    00000000  eb 52 90 4e 54 46 53 20  20 20 20 00 02 08 00 00  |.R.NTFS    .....|
    00000010  00 00 00 00 00 f8 00 00  00 00 00 00 00 00 00 00  |................|
                                                   ^^ ^^ ^^ ^^
Hidden Sectors value at offset 0x1C in the NTFS Partition Boot Sector.

So mkntfs is warning, writing the Hidden Sectors as zero and reporting
success.  Fix GParted in an equivalent way when it is updating the
Hidden Sectors for a moved or copied NTFS which starts at sector 2^32
and beyond.

After this fix the operational details for the same copy operation are:
    Copy /dev/loop0p1 to /dev/loop0p2
    * calibrate /dev/loop0p1                                   (SUCCESS)
    * calibrate /dev/loop0p2                                   (SUCCESS)
    * set partition type on /dev/loop0p2                       (SUCCESS)
    * copy file system from /dev/loop0p1 to /dev/loop0p2       (SUCCESS)
    * update boot sector of ntfs file system on /dev/loop0p2   (WARNING)
        Partition start (4294967296) is beyond sector 4294967295 (2^32-1).
        Windows will not be able to boot from this file system.
    * check file system on /dev/loop0p2 for errors and if p... (SUCCESS)

[1] NTFS, Partition Boot Sector (PBS)
    "
    Byte     Field     Field name       Purpose
    offset   length
    0x1C     5 bytes   Hidden Sectors   The number of sectors preceding
                                        the partition.
    "
    https://en.wikipedia.org/wiki/NTFS#Partition_Boot_Sector_(PBS)

[2] std::string::insert
    "fill (5)  string& insert(size_t pos, size_t n, char c);

    Insert into string
    Inserts additional characters into the string right before the
    character indicated by pos (or p):

    (5) fill
        Insert n consecutive copies of character c.
    "
    https://cplusplus.com/reference/string/string/insert/

Closes #164 - GParted crashes copying NTFS partition to starting beyond
              2TiB